TDR-1318 Send task heartbeat.
The step function has been updated to expect a task heartbeat every
minute or it will time the task out. This change will send one every
minute. It is run in a separate fiber so it won't block the thread.

This needs a change to AWS utils so the build will fail until that is
deployed and updated in the dependencies file.