Add extra logging to the sendHeartbeat method.
I'm not completely sure why but somewhere between the async thread this
runs in and the async response from the AWS SDK, any errors are ignored
in the main thread.
Calling attempt and then logging based on the value of the either will
log if there are any problems sending the heartbeat.