Refactor the export
The main class is getting quite long and it's about to get longer now
we're driving the metadata csv from the custom metadata query so I've
moved the BagMetadata into a different method.

I've also changed the graphql API to take a config and consignmentId as
arguments to the class. This avoids each method having to declare them
in the arguments.

I've updated the tests to match.