Close the BagValidator class.
The `BagValidator` class creates a new thread pool when it's
instantiated. There is a close method which you're supposed to call on
the class which terminates the thread pool but we're not doing that.

I've wrapped this up in a cats-effect `Resource` which is the standard
way of dealing with things that need closing and it's now exiting
immediately.