Add a new export process.
The new process copies the files from the clean bucket to a new export
bucket and then reads the metadata from the database and writes that to
a sidecar file.

The original export is unchanged apart from being moved to a sub module.

The release process now publishes a tar.gz file for each submodule.
These are released to GitHub together. When the image builds, it pulls
both export processes into the image.

When we come to run this, there is a seperate step in the step function
for each export process but they're running the same task but with a
different command. This should make it fairly easy to retire the old
process if we need to.

We're also sending SNS messages for each file copied over.