Add list objects pagination and SNS error checking.
I noticed that this was only sending 1000 messages to the SNS topic.

I originally thought that this was some kind of soft limit on the number
of messages so I've grouped them into batches of 500 and added a 1
second wait between them. It will also check for errors and try again if
there are any.

It turned out that this wasn't the problem after all.

The problem is that listObjectsV2 only returns a maximum of 1000 items.
I really should have noticed that. It now recursively fetches
everything.

Given that the error checking and grouping of the SNS send isn't a bad
idea, even though that wasn't actually the issue, I'm going to leave it
in.